+++ crates/service/src/middleware.rs
@@ -1,23 +1,23 @@
 // Copyright 2023-, Edge & Node, GraphOps, and Semiotic Labs.
 // SPDX-License-Identifier: Apache-2.0
 
+mod allocation;
 mod attestation;
+mod attestation_signer;
 pub mod auth;
-mod inject_allocation;
-mod inject_attestation_signer;
-mod inject_context;
-mod inject_deployment;
-mod inject_labels;
-mod inject_receipt;
-mod inject_sender;
+mod deployment;
+mod labels;
 mod prometheus_metrics;
+mod sender;
+mod tap_context;
+mod tap_receipt;
 
-pub use inject_allocation::{allocation_middleware, Allocation, AllocationState};
-pub use inject_context::context_middleware;
-pub use inject_deployment::deployment_middleware;
-pub use inject_labels::labels_middleware;
-pub use inject_receipt::receipt_middleware;
-pub use inject_sender::{sender_middleware, SenderState};
-pub use inject_attestation_signer::{signer_middleware, AttestationState};
+pub use allocation::{allocation_middleware, Allocation, AllocationState};
 pub use attestation::{attestation_middleware, AttestationInput};
+pub use attestation_signer::{signer_middleware, AttestationState};
+pub use deployment::deployment_middleware;
+pub use labels::labels_middleware;
 pub use prometheus_metrics::PrometheusMetricsMiddlewareLayer;
+pub use sender::{sender_middleware, SenderState};
+pub use tap_context::context_middleware;
+pub use tap_receipt::receipt_middleware;
