+++ crates/service/src/middleware/attestation_signer.rs
